Update Jackson to support Json-schema-validator
Json-schema-validator is a 3p library that depends on Jackson. It depends on version 2.17 and references some newer functions. The current version of Jackson in Android is 2.12.

This change adds the missing functions by copying them directly from latest source. These are minor utility functions that were straightforward to add in.

I'm unsure of the risk in updating Androids version of Jackson. I'm open to doing that instead if it is preferable/feasible.

Test: Built locally
Bug: b/372209928
Change-Id: I692d5d84158dad3dca16358672081fa50d562db0